Move line pragma extensions to target CodeRenderingContext
The BuildLinePragma and BuildEnhancedLinePragma extension methods target CodeWriter and take a CodeRenderingContext that must own the CodeWriter. Instead of forcing a strong dependency between the parameters, this change moves the extension methods to target CodeRenderingContext and removes the CodeWriter parameter.
