+++ src/Compiler/Microsoft.CodeAnalysis.Razor.Compiler/src/Language/Components/ComponentNodeWriter.cs
@@ -371,7 +371,7 @@ internal abstract class ComponentNodeWriter : IntermediateNodeWriter, ITemplateT
             {
                 var attributeSourceSpan = (SourceSpan)(attribute.PropertySpan ?? attribute.OriginalAttributeSpan);
                 var requiresEscaping = attribute.PropertyName.IdentifierRequiresEscaping();
-                using (context.CodeWriter.BuildEnhancedLinePragma(attributeSourceSpan, context, characterOffset: requiresEscaping ? 1 : 0))
+                using (context.BuildEnhancedLinePragma(attributeSourceSpan, characterOffset: requiresEscaping ? 1 : 0))
                 {
                     context.CodeWriter.WriteIdentifierEscapeIfNeeded(attribute.PropertyName);
                     context.CodeWriter.WriteLine(attribute.PropertyName);
