+++ src/Compiler/Microsoft.CodeAnalysis.Razor.Compiler/src/Language/Extensions/DefaultTagHelperTargetExtension.cs
@@ -377,7 +377,7 @@ internal sealed class DefaultTagHelperTargetExtension : IDefaultTagHelperTargetE
                 var firstMappedChild = node.Children.FirstOrDefault(child => child.Source != null) as IntermediateNode;
                 var valueStart = firstMappedChild?.Source;
 
-                using (context.CodeWriter.BuildLinePragma(node.Source, context))
+                using (context.BuildLinePragma(node.Source))
                 {
                     var accessor = GetPropertyAccessor(node);
                     var assignmentPrefixLength = accessor.Length + " = ".Length;
@@ -567,7 +567,7 @@ internal sealed class DefaultTagHelperTargetExtension : IDefaultTagHelperTargetE
             }
             else
             {
-                using (context.CodeWriter.BuildEnhancedLinePragma(token.Source, context))
+                using (context.BuildEnhancedLinePragma(token.Source))
                 {
                     context.CodeWriter.Write(token.Content);
                 }
