+++ src/Compiler/Microsoft.CodeAnalysis.Razor.Compiler/src/Language/Extensions/DesignTimeDirectiveTargetExtension.cs
@@ -74,7 +74,7 @@ internal class DesignTimeDirectiveTargetExtension : IDesignTimeDirectiveTargetEx
                     }
 
                     // {node.Content} __typeHelper = default({node.Content});
-                    using (context.CodeWriter.BuildLinePragma(node.Source, context))
+                    using (context.BuildLinePragma(node.Source))
                     {
                         context.AddSourceMappingFor(node);
                         context.CodeWriter
@@ -108,7 +108,7 @@ internal class DesignTimeDirectiveTargetExtension : IDesignTimeDirectiveTargetEx
                     }
 
                     // global::System.Object {node.content} = null;
-                    using (context.CodeWriter.BuildLinePragma(node.Source, context))
+                    using (context.BuildLinePragma(node.Source))
                     {
                         context.CodeWriter
                         .Write("global::")
@@ -140,7 +140,7 @@ internal class DesignTimeDirectiveTargetExtension : IDesignTimeDirectiveTargetEx
                     }
 
                     // global::System.Object __typeHelper = nameof({node.Content});
-                    using (context.CodeWriter.BuildLinePragma(node.Source, context))
+                    using (context.BuildLinePragma(node.Source))
                     {
                         context.CodeWriter
                         .Write("global::")
@@ -174,7 +174,7 @@ internal class DesignTimeDirectiveTargetExtension : IDesignTimeDirectiveTargetEx
                     }
 
                     // global::System.Object __typeHelper = "{node.Content}";
-                    using (context.CodeWriter.BuildLinePragma(node.Source, context))
+                    using (context.BuildLinePragma(node.Source))
                     {
                         context.CodeWriter
                         .Write("global::")
@@ -202,7 +202,7 @@ internal class DesignTimeDirectiveTargetExtension : IDesignTimeDirectiveTargetEx
 
                 case DirectiveTokenKind.Boolean:
                     // global::System.Boolean __typeHelper = {node.Content};
-                    using (context.CodeWriter.BuildLinePragma(node.Source, context))
+                    using (context.BuildLinePragma(node.Source))
                     {
                         context.CodeWriter
                         .Write("global::")
@@ -231,7 +231,7 @@ internal class DesignTimeDirectiveTargetExtension : IDesignTimeDirectiveTargetEx
         // for consistency so when a C# completion session starts, filling user code doesn't result in
         // a previously non-existent line pragma from being added and destroying the context in which
         // the completion session was started.
-        using (context.CodeWriter.BuildLinePragma(node.Source, context))
+        using (context.BuildLinePragma(node.Source))
         {
             context.AddSourceMappingFor(node);
             context.CodeWriter.Write(" ");
