+++ src/Compiler/Microsoft.CodeAnalysis.Razor.Compiler/src/Mvc/InjectTargetExtension.cs
@@ -30,7 +30,7 @@ public class InjectTargetExtension(bool considerNullabilityEnforcement) : IInjec
             if (node.TypeName == "")
             {
                 // if we don't even have a type name, just emit an empty mapped region so that intellisense still works
-                using (context.CodeWriter.BuildEnhancedLinePragma(node.TypeSource.Value, context))
+                using (context.BuildEnhancedLinePragma(node.TypeSource.Value))
                 {
                 }
             }
@@ -51,7 +51,7 @@ public class InjectTargetExtension(bool considerNullabilityEnforcement) : IInjec
 
             if (node.Source.HasValue)
             {
-                using (context.CodeWriter.BuildLinePragma(node.Source.Value, context))
+                using (context.BuildLinePragma(node.Source.Value))
                 {
                     WriteProperty();
                 }
